requirements.txt
# Core Agent Framework
google-adk>=1.19.0

# LLM Integration
litellm>=1.80.0

# Data & External Integrations
snowflake-snowpark-python>=1.11.0

# API Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Observability
opentelemetry-api>=1.22.0
opentelemetry-sdk>=1.22.0
opentelemetry-instrumentation-fastapi>=0.43b0
opentelemetry-exporter-gcp-trace>=1.6.0

# Security & Utils
python-dotenv>=1.0.0
python-multipart>=0.0.6

# Bonus Features
google-cloud-firestore>=2.14.0
prometheus-client>=0.19.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
requests>=2.31.0  # For integration tests